requirements.txt
Django==1.9
django-pagedown==0.1.1
djangorestframework==3.3.3
MySQL-python==1.2.5
